Work around a stupid linker error.
The error generated is:

    ld: unknown option: -no_deduplicate

I can't remember the exact cause but it's to do with some
workaround/compatibility issue implemented by LLVM (I think) but
which failed to account for older linker versions.

The other alternative is to use a different `-O` level--except you
can't use `-O0` but thanks to the wonder of *a different*
implementation bug you can use `-O00`. :D

I think the issue only happens on Mac and only on older versions.
(And I only know the workarounds because I've encountered this
before...)