Fix to not check error state when a valid handle was returned.
This fixes a (at the time) puzzling issue on Windows when certain
DLLs couldn't be opened in Godot even though they were actually
being opened by LoadLibrary.

The issue was that nothing was resetting the LastError value so
when we (mistakenly) checked the value sometimes it indicated an
error that wasn't caused by the DLL load operation.

We should only check for an error if an invalid handle is returned.